trailer_ai/requirements.txt
yt-dlp
opencv-python
librosa==0.10.1
numpy==1.24.4
pandas
tqdm
lightgbm
scikit-learn
joblib
python-dotenv
matplotlib
